Adjust Structure of Flux Plugin
